chore(file naming) remove duplicated files and rename files to use same naming convention - remove duplicated files - rename models files [Finishes #169939836]
