--- seeders/20191118205247-demo-booked-room.js
@@ -1,10 +1,0 @@
-module.exports = {
-  up: queryInterface => queryInterface.bulkInsert('bookedRooms', [{
-    bookingId: 1,
-    roomId: 1,
-    createdAt: '2019-12-02',
-    updatedAt: '2019-12-02',
-  }], {}),
-
-  down: queryInterface => queryInterface.bulkDelete('bookedRooms', null, {})
-};
