+++ src/models/Trip.js
