--- src/models/booking.js
@@ -1,22 +1,0 @@
-module.exports = (sequelize, DataTypes) => {
-  const Booking = sequelize.define('booking', {
-    userId: DataTypes.INTEGER,
-    hotelId: DataTypes.INTEGER,
-    roomId: DataTypes.INTEGER
-  }, {});
-  Booking.associate = (models) => {
-    Booking.belongsTo(models.user, {
-      foreignkey: 'userId',
-      onDelete: 'CASCADE'
-    });
-    Booking.belongsTo(models.hotel, {
-      foreignkey: 'hotelId',
-      onDelete: 'CASCADE'
-    });
-    Booking.belongsTo(models.hotel, {
-      foreignkey: 'roomId',
-      onDelete: 'CASCADE'
-    });
-  };
-  return Booking;
-};
